Do blocking work in blocking threads to avoid UI hanging on demanding workloads
