+++ src-tauri/src/main.rs
@@ -4,30 +4,40 @@
 )]
 use std::{collections::HashMap, fs::File, sync::Arc};
 
-use modmanager::{Mod, ModChangedEvent, ModManager};
+use lazy_static::lazy_static;
+use modmanager::{ModChangedEvent, ModManager};
 use simplelog::{
     ColorChoice, CombinedLogger, Config, LevelFilter, TermLogger, TerminalMode, WriteLogger,
 };
-use tauri::{Manager, State, Window};
+use tauri::{Manager, RunEvent, State, Window};
 
 mod config;
 mod db;
 mod error;
-mod filetree;
 mod modmanager;
 
-use config::ModloaderConfig;
+use config::{ConfigError, ModloaderConfig};
 use db::AppDb;
 use error::Result;
-
-use crate::config::ConfigError;
+use modmanager::Mod;
 
 const DB_PATH: &str = "db";
 const DB_FLUSH_INTERVAL: u64 = 500;
 const DB_CACHE_CAPACITY: u64 = 10_485_760;
 
+lazy_static! {
+    /// The database of the application
+    static ref DB: AppDb = {
+        AppDb::open(
+            &config::APP_SAVE_PATH.join(DB_PATH),
+            DB_FLUSH_INTERVAL,
+            DB_CACHE_CAPACITY,
+        )
+    };
+}
+
 fn main() {
-    tauri::Builder::default()
+    let app = tauri::Builder::default()
         .invoke_handler(tauri::generate_handler![
             set_advanced_config,
             config_is_initialized,
@@ -42,11 +52,6 @@ fn main() {
             activate_mod,
             deactivate_mod
         ])
-        .manage(Arc::new(AppDb::open(
-            &config::APP_SAVE_PATH.join(DB_PATH),
-            DB_FLUSH_INTERVAL,
-            DB_CACHE_CAPACITY,
-        )))
         .manage(
             ModManager::new()
                 .map_err(|e| log::error!("Failed to initialize ModManager: {:?}", e))
@@ -108,28 +113,33 @@ fn main() {
 
             Ok(())
         })
-        .run(tauri::generate_context!())
-        .expect("error while running tauri application");
+        .build(tauri::generate_context!())
+        .expect("error while building tauri application");
+
+    app.run(|_, event| {
+        if let RunEvent::Exit = event {
+            DB.flush()
+        }
+    });
 }
 
 // Configuration related commands
 /// Check if the config contains all required configuration parameters and is initialized
 #[tauri::command]
-fn config_is_initialized() -> Result<bool> {
-    let config = ModloaderConfig::load_config()?;
+async fn config_is_initialized() -> Result<bool> {
+    let config = ModloaderConfig::load_config().await?;
 
     Ok(config.get_game_path().is_some() && config.get_game_language().is_some())
 }
 
 /// Set the advanced configuration of the app (game language and game path)
 #[tauri::command]
-fn set_advanced_config(
+async fn set_advanced_config(
     game_lang: String,
     game_path: String,
-    db: State<'_, Arc<AppDb>>,
     mod_manager: State<'_, ModManager>,
 ) -> Result<()> {
-    let mut config = ModloaderConfig::load_config()?;
+    let mut config = ModloaderConfig::load_config().await?;
 
     let current_game_lang = config.get_game_language();
     let current_game_path = config.get_game_path();
@@ -143,58 +153,58 @@ fn set_advanced_config(
         _ => Err(ConfigError::GameLanguageNotSupported),
     }?;
 
-    if current_game_lang.is_some() && current_game_path.is_some() {
-        if current_game_lang.unwrap() == game_lang_cmp
-            && current_game_path.unwrap().to_string_lossy().to_string() == game_path
-        {
-            // Nothing changed in the configuration
-            return Ok(());
-        }
+    if current_game_lang.is_some()
+        && current_game_path.is_some()
+        && current_game_lang.unwrap() == game_lang_cmp
+        && current_game_path.unwrap().to_string_lossy() == game_path
+    {
+        // Nothing changed in the configuration
+        return Ok(());
     }
 
     // Deactivate all active mods as the config change requires a change of the mod installation folder inside the game installation
-    mod_manager.deactivate_all_mods(&db)?;
+    mod_manager.deactivate_all_mods().await?;
 
-    config.set_advanced_config(game_lang, game_path)?;
+    config.set_advanced_config(game_lang, game_path).await?;
 
     Ok(())
 }
 
 #[tauri::command]
-fn configure_dark_mode(dark: bool) -> Result<bool> {
-    let mut config = ModloaderConfig::load_config()?;
+async fn configure_dark_mode(dark: bool) -> Result<bool> {
+    let mut config = ModloaderConfig::load_config().await?;
 
-    config.set_dark_theme(dark)?;
+    config.set_dark_theme(dark).await?;
 
     Ok(dark)
 }
 
 #[tauri::command]
-fn get_dark_mode() -> Result<bool> {
-    let config = ModloaderConfig::load_config()?;
+async fn get_dark_mode() -> Result<bool> {
+    let config = ModloaderConfig::load_config().await?;
 
     Ok(config.get_dark_theme())
 }
 
 #[tauri::command]
-fn get_app_language() -> Result<Option<String>> {
-    let config = ModloaderConfig::load_config()?;
+async fn get_app_language() -> Result<Option<String>> {
+    let config = ModloaderConfig::load_config().await?;
 
     Ok(config.get_app_language())
 }
 
 #[tauri::command]
-fn set_app_language(lang: String) -> Result<()> {
-    let mut config = ModloaderConfig::load_config()?;
+async fn set_app_language(lang: String) -> Result<()> {
+    let mut config = ModloaderConfig::load_config().await?;
 
-    config.set_app_language(Some(lang))?;
+    config.set_app_language(Some(lang)).await?;
 
     Ok(())
 }
 
 #[tauri::command]
-fn get_advanced_config() -> Result<(Option<String>, Option<String>)> {
-    let config = ModloaderConfig::load_config()?;
+async fn get_advanced_config() -> Result<(Option<String>, Option<String>)> {
+    let config = ModloaderConfig::load_config().await?;
 
     Ok((
         config.get_game_language(),
@@ -206,11 +216,8 @@ fn get_advanced_config() -> Result<(Option<String>, Option<String>)> {
 
 // Mod related commands
 #[tauri::command]
-fn get_initial_mod_data(
-    mod_manager: State<'_, ModManager>,
-    db: State<'_, Arc<AppDb>>,
-) -> Result<HashMap<u64, Mod>> {
-    let data = mod_manager.get_initial_mod_data(&db)?;
+fn get_initial_mod_data(mod_manager: State<'_, ModManager>) -> Result<HashMap<u64, Mod>> {
+    let data = mod_manager.get_initial_mod_data()?;
 
     Ok(data)
 }
@@ -219,43 +226,30 @@ fn get_initial_mod_data(
 async fn add_new_mod(
     mod_manager: State<'_, ModManager>,
     window: Window,
-    db: State<'_, Arc<AppDb>>,
     archive_path: String,
 ) -> Result<()> {
-    mod_manager.add_mod(&archive_path, &db, window).await?;
+    mod_manager.add_mod(&archive_path, window).await?;
 
     Ok(())
 }
 
 #[tauri::command]
-fn delete_mod(
-    mod_manager: State<'_, ModManager>,
-    uid: u64,
-    db: State<'_, Arc<AppDb>>,
-) -> Result<()> {
-    mod_manager.delete_mod(uid, &db)?;
+async fn delete_mod(mod_manager: State<'_, ModManager>, uid: u64) -> Result<()> {
+    mod_manager.delete_mod(uid).await?;
 
     Ok(())
 }
 
 #[tauri::command]
-fn activate_mod(
-    mod_manager: State<'_, ModManager>,
-    uid: u64,
-    db: State<'_, Arc<AppDb>>,
-) -> Result<()> {
-    mod_manager.activate_mod(uid, &db)?;
+async fn activate_mod(mod_manager: State<'_, ModManager>, uid: u64) -> Result<()> {
+    mod_manager.activate_mod(uid).await?;
 
     Ok(())
 }
 
 #[tauri::command]
-fn deactivate_mod(
-    mod_manager: State<'_, ModManager>,
-    uid: u64,
-    db: State<'_, Arc<AppDb>>,
-) -> Result<()> {
-    mod_manager.deactivate_mod(uid, &db)?;
+async fn deactivate_mod(mod_manager: State<'_, ModManager>, uid: u64) -> Result<()> {
+    mod_manager.deactivate_mod(uid).await?;
 
     Ok(())
 }
