+++ src-tauri/src/modmanager/filetree.rs
@@ -1,13 +1,13 @@
 //! The file tree is used to keep track on which files have been modified/created by which mods.
 //! This allows to check whether mods collide with each other. And allows efficient file cleanup/restore on mod deactivation/deletion
-use std::fs::File;
 use std::vec;
 
 use serde::{Deserialize, Serialize};
 use sled::transaction::{TransactionError, UnabortableTransactionError};
 
-use crate::db::{AppDb, BincodeTransactional, Key};
-use crate::modmanager::{ModManagerError, Result};
+use super::error::{ModManagerError, Result};
+use crate::db::{BincodeTransactional, Key};
+use crate::DB;
 
 const DB_FILE_TREE_NAME: &str = "filetree";
 
@@ -20,11 +20,8 @@ pub struct FileTreeManager;
 impl FileTreeManager {
     /// Checks if the provided list of file paths does not conflict with any existing file tree entries.
     /// Returns None if no conflicts are found and a vector containing the conflicting mod uid's with the corresponding conflicting path.
-    pub fn get_conflicts(
-        file_paths: &Vec<String>,
-        db: &AppDb,
-    ) -> Result<Option<Vec<(u64, String)>>> {
-        let tree = db.open_tree(DB_FILE_TREE_NAME);
+    pub fn get_conflicts(file_paths: &Vec<String>) -> Result<Option<Vec<(u64, String)>>> {
+        let tree = DB.open_tree(DB_FILE_TREE_NAME);
 
         let conflicts = tree
             .transaction::<_, _, UnabortableTransactionError>(|transaction| {
@@ -57,8 +54,8 @@ impl FileTreeManager {
     }
 
     /// Insert the list of file paths into the file tree with the corresponding mod uid
-    pub fn insert_files(uid: u64, file_paths: &Vec<String>, db: &AppDb) -> Result<()> {
-        let tree = db.open_tree(DB_FILE_TREE_NAME);
+    pub fn insert_files(uid: u64, file_paths: &Vec<String>) -> Result<()> {
+        let tree = DB.open_tree(DB_FILE_TREE_NAME);
 
         tree.transaction::<_, _, UnabortableTransactionError>(|transaction| {
             for file_path in file_paths {
@@ -80,8 +77,8 @@ impl FileTreeManager {
     }
 
     /// Get all files in the tree owned by the provided mod uid
-    pub fn get_files(uid: u64, db: &AppDb) -> Result<Vec<String>> {
-        let tree = db.open_tree(DB_FILE_TREE_NAME);
+    pub fn get_files(uid: u64) -> Result<Vec<String>> {
+        let tree = DB.open_tree(DB_FILE_TREE_NAME);
 
         let mut found_paths = vec![];
 
@@ -102,12 +99,12 @@ impl FileTreeManager {
             found_paths.push(path);
         }
 
-        return Ok(found_paths);
+        Ok(found_paths)
     }
 
     /// Remove all provided file paths from the tree
-    pub fn remove_files(paths: &Vec<String>, db: &AppDb) -> Result<()> {
-        let tree = db.open_tree(DB_FILE_TREE_NAME);
+    pub fn remove_files(paths: &Vec<String>) -> Result<()> {
+        let tree = DB.open_tree(DB_FILE_TREE_NAME);
 
         tree.transaction::<_, _, UnabortableTransactionError>(|transaction| {
             for file_path in paths {
@@ -127,30 +124,4 @@ impl FileTreeManager {
 
         Ok(())
     }
-}
-
-/// Gets all paths of the files and dirs contained in an archive. Automatically removes modinfo.json path if it exists as it does not need to be injected into the game.
-pub fn get_archive_dirs_and_files(archive: &File) -> Result<(Vec<String>, Vec<String>)> {
-    let mut files = compress_tools::list_archive_files(archive)?;
-    let mut dirs = vec![];
-
-    let mut idx = 0;
-
-    while idx < files.len() {
-        let path = &files[idx];
-
-        if path.ends_with('/') {
-            dirs.push(files.remove(idx));
-            continue;
-        }
-
-        if path.contains("modinfo.json") {
-            files.remove(idx);
-            continue;
-        }
-
-        idx += 1;
-    }
-
-    Ok((files, dirs))
 }
